backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
fastapi==0.115.0
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==1.26.4
pandas==2.2.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
six==1.17.0
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.30.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
